Refactor the deployment source into a separate file.
